fix #855: enable uploads of files > 2G to NFS

+++ mucommander-protocol-nfs/src/main/java/com/sun/nfs/Nfs.java
@@ -476,7 +476,7 @@ public abstract class Nfs {
              */
             Buffer b = bufferList[index];
             if (b == null) {
-                b = new Buffer(this, index * wsize, wsize);
+                b = new Buffer(this, (long) index * wsize, wsize);
                 bufferList[index] = b;
             }
